:bug: fix Crc64Digest declaration

+++ Bell/Hash/Crc64.hpp
@@ -14,7 +14,7 @@
 
 namespace Bell { namespace Hash { namespace Crc64 {
 
-	using Crc64Digest = std::uint32_t;
+	using Crc64Digest = std::uint64_t;
 
 	/**
 	 * @brief      CRC-64
